requirements.txt
Django==4.2.4
django-precise-bbcode==1.2.16
Pillow